+++ HISTORY.rst
@@ -6,6 +6,8 @@ Bug tracker at https://github.com/giampaolo/psutil/issues
 **Enhancements**
 
 - #581: add .gitignore. (patch by Gabi Davar)
+- #582: connection constants returned by psutil.net_connections() and
+  psutil.Process.connections() were turned from int to enums on Python > 3.4.
 
 
 2.2.1 - 2015-02-02